requirements.txt
langchain==0.1.16
langchain-community==0.0.32
langchain-core==0.1.42
langchain-mistralai==0.1.2
langchain-openai==0.0.8
langchain-text-splitters==0.0.1
langdetect==1.0.9
langsmith==0.1.45
PyQt5==5.15.10
PyQt5_sip==12.13.0
pytest==7.4.0
python-dotenv==1.0.1
Requests==2.31.0
unstructured==0.13.2
unstructured-client==0.18.0
openai==1.17.0
chroma-hnswlib==0.7.3
chromadb==0.4.24
tiktoken==0.5.2
zipp==3.17.0
